requirements.txt
asgiref==3.8.1
Django==5.1
pillow==10.4.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
sqlparse==0.5.1
